Fix newsletter form and search size
