--- src/components/Newsletter/NewsletterForm.js
@@ -1,65 +1,0 @@
-import React from 'react'
-import PropTypes from 'prop-types'
-import { translate } from 'react-i18next'
-
-import { getUser } from '../../fetch/fetch'
-import { waitForDataAndSetState, cancelAllPromises } from '../../helpers/components'
-import Button from '../Buttons/Button'
-
-import style from './NewsletterForm.scss'
-
-class NewsletterForm extends React.Component {
-  static propTypes = {
-    t: PropTypes.func.isRequired
-  }
-
-  constructor(props) {
-    super(props)
-    this.state = {
-      user: null,
-      value: '',
-      errors: []
-    }
-  }
-
-  componentDidMount() {
-    return waitForDataAndSetState(getUser(), this, 'user')
-      .then(() => {
-        if (!this.state.user) return
-        this.setState({ value: this.state.user.email })
-      })
-  }
-
-  componentWillUnmount() {
-    return cancelAllPromises(this)
-  }
-
-  handleChange = event => {
-    this.setState({ value: event.target.value })
-  }
-
-  render() {
-    const { t } = this.props
-    const { value } = this.state
-
-    return (
-      <div className={style.container}>
-        <form action='//gouv.us15.list-manage.com/subscribe/post?u=f4e80584578b65fde5aadffb6&amp;id=a9e2a3104d' method='post'name='mc-embedded-subscribe-form' target='_blank' noValidate>
-          <h2>{t('components.NewsletterForm.newsletterSignUp')}</h2>
-          <input type='email' name='EMAIL' placeholder={t('components.NewsletterForm.label')} value={value} />
-
-          <Button type='submit' text={t('components.NewsletterForm.inputValue')} name='subscribe' style={{
-            padding: '0.85em 3em',
-            marginTop: '1em',
-            width: '100%',
-            textTransform: 'uppercase',
-            fontSize: '1.2em',
-            fontWeight: 'bold'
-          }} />
-        </form>
-      </div>
-    )
-  }
-}
-
-export default translate('Common')(NewsletterForm)
