+++ src/components/NewsletterForm/NewsletterForm.scss
@@ -19,14 +19,7 @@
   }
 }
 
-form {
+.form {
   max-width: 640px;
   margin: auto;
-}
-
-footer {
-  h2 {
-    margin-top: 0;
-    margin-bottom: 0.5em;
-  }
 }
