+++ src/components/NewsletterForm/index.js
@@ -1,0 +1,3 @@
+import NewsletterForm from './NewsletterForm'
+
+export default NewsletterForm
